requirements/develop.txt
-r production.txt
bpython
django-cors-headers
django-debug-toolbar